Revert "tutorial work"
This reverts commit 10c868ccd3cb184813d39aaacbddf743fd595d37, reversing
changes made to 7a0416a388434bea7ec585554ee397e948f51cb3.

+++ Assets/Scripts/AttackFireball.cs
@@ -9,7 +9,7 @@ public class AttackFireball : BaseAttack
     {
         speed = 15;
         damage = 7;
-        element = Element.Fire;
+        element = (int) Element.Fire;
     }
 
     // Update is called once per frame
